requirements.txt
requests
python3-bs4
html5lib
lxml
selenium